# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlowCycle - Smart Period Tracker & Women's Health Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  age: number;
  heightCm: number;
  weightKg: number;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  age: number;
  heightCm: number;
  weightKg: number;
  bmi: number;
  bmiCategory: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface UpdateProfileBody {
  name?: string;
  age?: number;
  heightCm?: number;
  weightKg?: number;
}

export interface Cycle {
  id: number;
  userId: number;
  startDate: string;
  cycleLength: number;
  notes?: string;
  createdAt: string;
}

export interface CreateCycleBody {
  startDate: string;
  cycleLength: number;
  notes?: string;
}

export interface CyclePrediction {
  lastPeriodStart: string;
  nextPeriodDate: string;
  ovulationDate: string;
  fertileWindowStart: string;
  fertileWindowEnd: string;
  currentPhase: string;
  cycleDay: number;
  cycleLength: number;
  daysUntilNextPeriod: number;
}

export interface CalendarNote {
  id: number;
  userId: number;
  date: string;
  note: string;
  createdAt: string;
}

export interface CreateCalendarNoteBody {
  date: string;
  note: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl?: string;
  inStock: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  category: string;
  imageUrl?: string;
  inStock: boolean;
}

export interface CartItem {
  id: number;
  productId: number;
  product: Product;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
}

export interface CheckoutResponse {
  orderId: string;
  message: string;
  total: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  age: number;
  isAdmin: boolean;
  cycleCount: number;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalCycles: number;
  totalProducts: number;
  activeUsersThisMonth: number;
}

export interface WellnessTips {
  phase: string;
  phaseDescription: string;
  dietTips: string[];
  exerciseTips: string[];
  selfCareTips: string[];
  moodTips: string[];
}

export interface Quote {
  text: string;
  author: string;
}

export type GetProductsParams = {
  category?: string;
};

export type GetWellnessTipsParams = {
  phase?: GetWellnessTipsPhase;
};

export type GetWellnessTipsPhase =
  (typeof GetWellnessTipsPhase)[keyof typeof GetWellnessTipsPhase];

export const GetWellnessTipsPhase = {
  menstrual: "menstrual",
  follicular: "follicular",
  ovulation: "ovulation",
  luteal: "luteal",
} as const;