# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlowCycle - Smart Period Tracker & Women's Health Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /cycles:
    get:
      operationId: getCycles
      tags: [cycles]
      summary: Get all cycle history for current user
      responses:
        "200":
          description: List of cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cycle"
    post:
      operationId: createCycle
      tags: [cycles]
      summary: Log a new cycle entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleBody"
      responses:
        "201":
          description: Cycle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"
...
  /cycles/{cycleId}:
    delete:
      operationId: deleteCycle
      tags: [cycles]
      summary: Delete a cycle entry
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /calendar/notes:
    get:
      operationId: getCalendarNotes
      tags: [calendar]
      summary: Get all calendar notes for current user
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarNote"
    post:
      operationId: createCalendarNote
      tags: [calendar]
      summary: Create or update a note for a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarNoteBody"
      responses:
        "201":
          description: Note saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarNote"
  /calendar/notes/{noteId}:
    delete:
      operationId: deleteCalendarNote
      tags: [calendar]
      summary: Delete a calendar note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user's cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear the cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/{itemId}:
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove an item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /admin/users/{userId}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin - Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        age:
          type: integer
        heightCm:
          type: number
        weightKg:
          type: number
      required: [name, email, password, age, heightCm, weightKg]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: integer
        heightCm:
          type: number
        weightKg:
          type: number
        bmi:
          type: number
        bmiCategory:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, age, heightCm, weightKg, bmi, bmiCategory, isAdmin, createdAt]
...
    Cycle:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        startDate:
          type: string
        cycleLength:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
      required: [id, userId, startDate, cycleLength, createdAt]
    CreateCycleBody:
      type: object
      properties:
        startDate:
          type: string
        cycleLength:
          type: integer
        notes:
          type: string
      required: [startDate, cycleLength]
...
    CalendarNote:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        note:
          type: string
        createdAt:
          type: string
      required: [id, userId, date, note, createdAt]
    CreateCalendarNoteBody:
      type: object
      properties:
        date:
          type: string
        note:
          type: string
      required: [date, note]
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        createdAt:
          type: string
      required: [id, name, description, price, category, inStock, createdAt]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
      required: [id, productId, product, quantity]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
      required: [items, total]
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
...