# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlowCycle - Smart Period Tracker & Women's Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  age: number;
  heightCm: number;
  weightKg: number;
}