# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlowCycle - Smart Period Tracker & Women's Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  age: number;
  heightCm: number;
  weightKg: number;
  bmi: number;
  bmiCategory: string;
  isAdmin: boolean;
  createdAt: string;
}